apps/roboflowapp/imos-roboflowapp/requirements.txt
flask
flask-cors
supervision
inference-sdk
Pillow